+++ subprojects/workers/src/main/java/org/gradle/workers/internal/ClassLoaderStructureProvider.java
@@ -16,12 +16,18 @@
 
 package org.gradle.workers.internal;
 
+import com.google.common.collect.Sets;
 import org.gradle.initialization.ClassLoaderRegistry;
 import org.gradle.initialization.MixInLegacyTypesClassLoader;
+import org.gradle.internal.classloader.ClasspathUtil;
 import org.gradle.internal.classloader.FilteringClassLoader;
 import org.gradle.internal.classloader.VisitableURLClassLoader;
+import org.gradle.internal.classpath.DefaultClassPath;
 
 import java.io.File;
+import java.net.URL;
+import java.util.ArrayList;
+import java.util.Set;
 
 public class ClassLoaderStructureProvider {
     private final ClassLoaderRegistry classLoaderRegistry;
@@ -33,7 +39,7 @@ public class ClassLoaderStructureProvider {
     public ClassLoaderStructure getWorkerProcessClassLoaderStructure(final Iterable<File> additionalClasspath, Class<?>... classes) {
         MixInLegacyTypesClassLoader.Spec workerExtensionSpec = classLoaderRegistry.getGradleWorkerExtensionSpec();
         FilteringClassLoader.Spec gradleApiFilter = classLoaderRegistry.getGradleApiFilterSpec();
-        VisitableURLClassLoader.Spec userSpec = classLoaderRegistry.getUserSpec("worker-loader", additionalClasspath, classes);
+        VisitableURLClassLoader.Spec userSpec = getUserSpec("worker-loader", additionalClasspath, classes);
 
         // Add the Gradle API filter between the user classloader and the worker infrastructure classloader
         return new HierarchicalClassLoaderStructure(workerExtensionSpec)
@@ -43,9 +49,30 @@ public class ClassLoaderStructureProvider {
 
     public ClassLoaderStructure getInProcessClassLoaderStructure(final Iterable<File> additionalClasspath, Class<?>... classes) {
         FilteringClassLoader.Spec gradleApiFilter = classLoaderRegistry.getGradleApiFilterSpec();
-        VisitableURLClassLoader.Spec userSpec = classLoaderRegistry.getUserSpec("worker-loader", additionalClasspath, classes);
+        VisitableURLClassLoader.Spec userSpec = getUserSpec("worker-loader", additionalClasspath, classes);
         // Add the Gradle API filter between the user classloader and the worker infrastructure classloader
         return new HierarchicalClassLoaderStructure(gradleApiFilter)
                 .withChild(userSpec);
+    }
+
+    /**
+     * Returns a spec representing the combined "user" classloader for the given classes and additional classpath.  The user classloader assumes it is used as a child of a classloader with the Gradle API.
+     */
+    public VisitableURLClassLoader.Spec getUserSpec(String name, Iterable<File> additionalClasspath, Class<?>... classes) {
+        Set<URL> classpath = Sets.newLinkedHashSet();
+        classpath.addAll(DefaultClassPath.of(additionalClasspath).getAsURLs());
+
+        Set<ClassLoader> uniqueClassloaders = Sets.newHashSet();
+        for (Class clazz : classes) {
+            ClassLoader classLoader = clazz.getClassLoader();
+            // System types come from the system classloader and their classloader is null.
+            if (classLoader != null) {
+                uniqueClassloaders.add(classLoader);
+            }
+        }
+        for (ClassLoader classLoader : uniqueClassloaders) {
+            ClasspathUtil.collectClasspathUntil(classLoader, classLoaderRegistry.getGradleApiClassLoader(), classpath);
+        }
+        return new VisitableURLClassLoader.Spec(name, new ArrayList<>(classpath));
     }
 }
